requirements.txt
loguru~=0.5.3
aiohttp~=3.8.6
pyyaml~=6.0.1
openai~=1.3.7
setuptools~=68.0.0
dynaconf~=3.2.4